+++ dbms/src/Debug/dbgFuncCoprocessorUtils.h
@@ -17,14 +17,11 @@
 #include <DataStreams/BlocksListBlockInputStream.h>
 #include <DataStreams/SquashingBlockInputStream.h>
 #include <Debug/DAGProperties.h>
-#include <Debug/dbgTools.h>
 #include <Flash/Coprocessor/ArrowChunkCodec.h>
 #include <Flash/Coprocessor/CHBlockChunkCodec.h>
 #include <Flash/Coprocessor/ChunkCodec.h>
 #include <Flash/Coprocessor/DAGContext.h>
-#include <Flash/Coprocessor/DAGDriver.h>
 #include <Flash/Coprocessor/DefaultChunkCodec.h>
-#include <Flash/CoprocessorHandler.h>
 #include <Interpreters/Context.h>
 #include <Interpreters/sortBlock.h>
 #include <Poco/StringTokenizer.h>
@@ -48,14 +45,4 @@ BlockInputStreamPtr outputDAGResponse(Context &, const DAGSchema & schema, const
 Block getMergedBigBlockFromDagRsp(Context & context, const DAGSchema & schema, const tipb::SelectResponse & dag_response);
 bool dagRspEqual(Context & context, const tipb::SelectResponse & expected, const tipb::SelectResponse & actual, String & unequal_msg);
 
-tipb::SelectResponse executeDAGRequest(
-    Context & context,
-    const tipb::DAGRequest & dag_request,
-    RegionID region_id,
-    UInt64 region_version,
-    UInt64 region_conf_version,
-    Timestamp start_ts,
-    std::vector<std::pair<DecodedTiKVKeyPtr, DecodedTiKVKeyPtr>> & key_ranges);
-
-bool runAndCompareDagReq(const coprocessor::Request & req, const coprocessor::Response & res, Context & context, String & unequal_msg);
 } // namespace DB
